requirements.txt
absl-py==1.0.0
apache-beam==2.38.0
appnope @ file:///opt/concourse/worker/volumes/live/4f734db2-9ca8-4d8b-5b29-6ca15b4b4772/volume/appnope_1606859466979/work
astunparse==1.6.3
avro-python3==1.10.2
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
debugpy @ file:///opt/concourse/worker/volumes/live/32b11d06-4d64-4ec8-497a-cf4fc97343d2/volume/debugpy_1637091821874/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.1.1
dm-tree==0.1.7
docopt==0.6.2
entrypoints @ file:///opt/concourse/worker/volumes/live/194c0a28-55ce-4e83-6a87-0d9f2e06ab2c/volume/entrypoints_1649926487944/work
fastavro==1.4.11
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gin-config==0.5.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
h5py==3.6.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel @ file:///opt/concourse/worker/volumes/live/c21734a4-132a-471d-7fe9-76688d5488be/volume/ipykernel_1647000841085/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/333da4fb-3db0-439c-4a1b-ee802c44da4f/volume/ipython_1643818165055/work
itsdangerous==2.1.2
jedi @ file:///opt/concourse/worker/volumes/live/c9d2fa99-8bc1-4572-41e7-6beba6391441/volume/jedi_1644315238822/work
Jinja2==3.1.2
joblib==1.1.0
jupyter-client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4d0a367d-74c5-447f-831e-3e1fd3984797bfrg56pj/croots/recipe/jupyter_client_1650622206454/work
jupyter-core @ file:///opt/concourse/worker/volumes/live/b49dfe5a-aa7f-4da7-60c3-fdd95ee78394/volume/jupyter_core_1646994453026/work
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.2.0
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
nest-asyncio @ file:///opt/concourse/worker/volumes/live/bf4e1da0-0622-4bab-4608-fdb487494886/volume/nest-asyncio_1649847922626/work
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
object-detection @ file:///Users/saitoinosaka/Documents/tensorflow2-movilenet2/TFODCourse/Tensorflow/models/research
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.1.2.30
opt-einsum==3.3.0
orjson==3.6.8
packaging==20.9
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
plotly==5.8.2
portalocker==2.4.0
progressbar2==4.0.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pymongo==3.12.3
pyparsing==2.4.7
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-slugify==6.1.2
python-utils==3.1.0
pytz==2022.1
PyYAML==5.4.1
pyzmq @ file:///opt/concourse/worker/volumes/live/7dec6841-64b1-4d95-7226-ce37c6ca8915/volume/pyzmq_1638435000749/work
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-io==0.25.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.2
tensorflowjs==3.18.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tornado @ file:///opt/concourse/worker/volumes/live/d531d395-893c-4ca1-6a5f-717b318eb08c/volume/tornado_1606942307627/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typeguard==2.13.3
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.1.2
wrapt==1.14.0
zipp==3.8.0
